Move debounce to test utils
This appeases Knip and indicates a little better that it's a test
function.
